refactor: move upgrade code to new file
